Refactor daos
