+++ libs/horizon/src/main/java/com/instructure/horizon/database/entity/HorizonDashboardProgramCourseRef.kt
@@ -13,17 +13,16 @@
  *     See the License for the specific language governing permissions and
  *     limitations under the License.
  */
-package com.instructure.horizon.database.sync
+package com.instructure.horizon.database.entity
 
 import androidx.room.Entity
-import androidx.room.PrimaryKey
 
-@Entity(tableName = "horizon_sync_metadata")
-data class HorizonSyncMetadataEntity(
-    @PrimaryKey val key: String,
-    val lastSyncedAtMs: Long,
-) {
-    companion object {
-        const val KEY_DASHBOARD_COURSES = "dashboard_courses"
-    }
-}
+@Entity(
+    tableName = "horizon_dashboard_program_course_refs",
+    primaryKeys = ["programId", "courseId"]
+)
+data class HorizonDashboardProgramCourseRef(
+    val programId: String,
+    val courseId: Long,
+    val enrollmentStatus: String?,
+)
